backend-deploy-UNUSED/requirements-render.txt
# Render.com Compatible Dependencies with PyMuPDF
requests==2.31.0
python-docx==1.1.0
PyMuPDF==1.24.2
pdfplumber==0.11.0
supabase==2.4.2
flask==3.0.2
gunicorn==21.2.0
flask-cors==4.0.0

# NLP libraries for job analysis
nltk==3.8.1
textblob==0.17.1

# Memory monitoring (optional - will fallback if not available)
psutil==5.9.8

# Gemini AI integration for CV optimization
google-generativeai==0.3.2